[SQUASH]: power toggle torch base: Allow doubletap/longpress power to toggle torch [1/2]
AEX Edits:
- Port to S
- handle camera gesture here instead of extensions
- simplify code

Thanks to beanstown106 for the initial longpress action calls in
PhoneWindowManager (improved by lineage guys)

If the user enables double tap for camera action, torch action will
automatically switch to long press mode.
If the user enables double tap for torch, double tap for camera
will be disabled accordingly.

Remove the settings pref is the device doesn't have flashlight

A 170ms delay for double tap action is not annoying on single tap
to wake up the device.

Vibrate when toggling torch to let the user know of unwanted toggling
in the pocket (even if vibration is disabled by user)

Allow torch action also on ambient display

mydongistiny: Fixed for android10
Change-Id: I12da044f86c7b625872607529cf8524615cf576b
Signed-off-by: dlwlrma123 <alexfinhart@gmail.com>

base: ScreenOff torch: Allow to disable torch when screen is on

Change-Id: I253815c7f1f988c20f4a8f041f9c66a811934b6f
Signed-off-by: DennySPb <dennyspb@gmail.com>
Co-authored-by: DennySPb <dennyspb@gmail.com>
Signed-off-by: dlwlrma123 <alexfinhart@gmail.com>